Avoid FP32 pre-cast for norm1 to enable AITER BF16 layernorm
WanTransformerBlock and WanTransformerBlock_VSA both did:
  norm1 = self.norm1(hidden_states.float())

The explicit .float() pre-cast forces FP32LayerNorm to take the slow
F.layer_norm FP32 path even on AMD ROCm (the BF16 AITER path requires
BF16 input).  Removing it lets the AITER-accelerated FP32LayerNorm
operate natively in BF16 (~5x faster).

Type correctness: scale_msa/shift_msa are FP32, so the multiply
  norm1 * (1 + scale_msa) + shift_msa
upcasts norm1 from BF16 to FP32 automatically via PyTorch type
promotion.  The result is cast to orig_dtype (BF16) as before.

Expected savings: ~75 ms per request
  (1 call/block × 0.250 ms/call × 300 blocks)

Combined with the FP32LayerNorm AITER commit (~152 ms), total layernorm
savings on ROCm are ~227 ms per 5-step Wan2.1 T2V-1.3B request.

+++ python/sglang/multimodal_gen/runtime/models/dits/wanvideo.py
@@ -382,7 +382,10 @@ class WanTransformerBlock(nn.Module):
         assert shift_msa.dtype == torch.float32
 
         # 1. Self-attention
-        norm1 = self.norm1(hidden_states.float())
+        # Pass hidden_states directly (without .float()) so FP32LayerNorm can
+        # use AITER BF16 layernorm on AMD ROCm (~5x faster than FP32 path).
+        # scale_msa/shift_msa are FP32, so type promotion gives FP32 product.
+        norm1 = self.norm1(hidden_states)
         norm_hidden_states = (norm1 * (1 + scale_msa) + shift_msa).to(orig_dtype)
         query, _ = self.to_q(norm_hidden_states)
         key, _ = self.to_k(norm_hidden_states)
@@ -547,8 +550,11 @@ class WanTransformerBlock_VSA(nn.Module):
         assert shift_msa.dtype == torch.float32
 
         # 1. Self-attention
+        # Pass hidden_states directly (without .float()) so FP32LayerNorm can
+        # use AITER BF16 layernorm on AMD ROCm (~5x faster than FP32 path).
+        # scale_msa/shift_msa are FP32, so type promotion gives FP32 product.
         norm_hidden_states = (
-            self.norm1(hidden_states.float()) * (1 + scale_msa) + shift_msa
+            self.norm1(hidden_states) * (1 + scale_msa) + shift_msa
         ).to(orig_dtype)
         query, _ = self.to_q(norm_hidden_states)
         key, _ = self.to_k(norm_hidden_states)
